[TestRunner] Implement Subnet Naming Convention
